Added Event System
Event system has 3 parts:
EventContainer: Holds lists of Actions that will be executed when all Conditions are completed.

EventCondition: Tracks other scripts using UnityEvents and delegates. When a condition is met, it alerts all listening EventContainers.

EventAction: Describes a set of actions to take place once all EventConditions in an EventContianer have been completed. Actions such as changing a material, enabling an object, teleporting the player, or interacting with a different script.

An example can be found in GameScene. Shooting all candles in the scene will replace the tables material with the flame material.