Cleaned up version
Cleaned up map and changed rendering settings to get better FPS. Still not perfect but it should be playable (>30fps) on a Quest 2. There's a FPS display on the back of the map to verify. This game is probably better off staying as PCVR due to performance issues.

Current game can be 'won', by bringing the artifact back to the sub and placing it inside. Several bugs fixed (such as dropping keys through the floor, crashes, etc).

I'll add in the enemies next if I get around to it